Move ont software validator to unaligned reads. Add tests
