+++ src/encoded/tests/data/workbook-inserts/supplementary_file.json
@@ -34,6 +34,9 @@
         ],
         "filename": "test_hela.fasta",
         "file_format": "FASTA",
+        "file_sets": [
+            "TEST_FILE-SET_HELA-HEK293-DNA"
+        ],
         "submission_centers": [
             "smaht"
         ],
