+++ src/encoded/types/unaligned_reads.py
@@ -1,4 +1,5 @@
 from typing import Union, List
+from pyramid.request import Request
 from snovault import collection, load_schema
 from snovault.util import debug_log, get_item_or_none
 from encoded.validator_decorators import link_related_validator
@@ -20,8 +21,15 @@ from .base import (
 from .utils import get_properties, get_property_for_validation
 from ..item_utils import (
      unaligned_reads as ur_utils,
-     file as file_utils
+     file as file_utils,
+     file_set as file_set_utils,
+     sequencer as sequencer_utils,
+     sequencing as sequencing_utils,
 )
+
+SEQUENCER_DEPENDENT = {
+    "ONT": ["gpu", "model", "modification_tags"]
+}
 
 @collection(
     name="unaligned-reads",
@@ -120,10 +128,68 @@ def check_read_pairs_in_file_sets(request, submitted_id: str, paired_with: Union
             return request.errors.add('body', 'UnalignedReads: invalid submission order', msg)
     return request.validated.update({})
 
+@link_related_validator
+def validate_basecalling_software_for_ont_on_add(context, request):
+    """Validate software and sequencer.platform if ONT on add"""
+    data = request.json
+    software = data['software'] if 'software' in data else None
+    file_sets = data['file_sets'] if 'file_sets' in data else None
+    return check_basecalling_software_for_ont(request, data['submitted_id'], software, file_sets)
+
+
+@link_related_validator
+def validate_basecalling_software_for_ont_on_edit(context, request):
+    """Validate software and sequencer.platform if ONT on edit."""
+    existing_properties = get_properties(context)
+    properties_to_update = get_properties(request)
+    software = get_property_for_validation('software', existing_properties, properties_to_update)
+    file_sets = get_property_for_validation('file_sets', existing_properties, properties_to_update)
+    return check_basecalling_software_for_ont(request, existing_properties['submitted_id'], software, file_sets)
+
+
+def check_basecalling_software_for_ont(
+    request: Request,
+    submitted_id: str,
+    software: Union[List[str], None],
+    file_sets: Union[List[str], None]
+):
+    """Check if the basecalling-specific properties for software are present if the sequencer.platform is ONT.
+    
+    Sequencer-specific properties are in `SEQUENCER_DEPENDENT`
+    """
+    if file_sets:
+        sequencings = [file_set_utils.get_sequencing(get_item_or_none(request, file_set, 'file_set')) for file_set in file_sets]
+        sequencer_aids = [sequencing_utils.get_sequencer(
+            get_item_or_none(request, sequencing, 'sequencing')
+        ) for sequencing in sequencings]
+        platforms = [sequencer_utils.get_platform(
+            get_item_or_none(request, sequencer_aid, 'sequencer')
+        ) for sequencer_aid in sequencer_aids]
+        for key, value in SEQUENCER_DEPENDENT.items():
+            if key in platforms:
+                # if not software:
+                #     msg = f"No software found for file {submitted_id}. Software items are required for {key} files."
+                #     return request.errors.add('body', 'SubmittedFile: invalid links', msg)
+                # else:
+                if software:
+                    sequencer_properties_present = False
+                    for sw in software:
+                        software_item = get_item_or_none(request, sw ,'software')
+                        if all(item in software_item.keys() for item in value):
+                            sequencer_properties_present = True
+                    if sequencer_properties_present:
+                        return request.validated.update({})
+                    else:
+                        msg = f"Sequencer-specific Software properties not found for file {submitted_id}. The following Software properties are required for {key} files: {value}."
+                        return request.errors.add('body', 'Software: invalid properties', msg)
+            return request.validated.update({})
+
 
 UNALIGNED_READS_ADD_VALIDATORS = SUBMITTED_FILE_ADD_VALIDATORS + [
     validate_read_pairs_on_add,
-    validate_read_pairs_in_file_sets_on_add
+    validate_read_pairs_in_file_sets_on_add,
+    validate_basecalling_software_for_ont_on_add,
+
 ]
 
 @view_config(
@@ -139,12 +205,14 @@ def unaligned_reads_add(context, request, render=None):
 
 UNALIGNED_READS_EDIT_PATCH_VALIDATORS = SUBMITTED_FILE_EDIT_PATCH_VALIDATORS + [
     validate_read_pairs_on_edit,
-    validate_read_pairs_in_file_sets_on_edit
+    validate_read_pairs_in_file_sets_on_edit,
+    validate_basecalling_software_for_ont_on_edit,
 ]
 
 UNALIGNED_READS_EDIT_PUT_VALIDATORS = SUBMITTED_FILE_EDIT_PUT_VALIDATORS + [
     validate_read_pairs_on_edit,
-    validate_read_pairs_in_file_sets_on_edit
+    validate_read_pairs_in_file_sets_on_edit,
+    validate_basecalling_software_for_ont_on_edit,
 ]
 
 @view_config(
